refactor script to package android/jvm using zip, and js/native using tar
