+++ library/kmp-tor-binary-extract/src/jsMain/kotlin/io/matthewnelson/kmp/tor/binary/extract/internal/-PlatformConstants.kt
@@ -17,25 +17,23 @@ package io.matthewnelson.kmp.tor.binary.extract.internal
 
 @Suppress("ObjectPropertyName", "SpellCheckingInspection")
 internal actual object PlatformConstants {
-    // TODO: Update with gunzipped archive details
-    private const val _ARCHIVE_SHA256_VALUE_GEOIP = ""
+    private const val _ARCHIVE_SHA256_VALUE_GEOIP = "d25e55d538b0628a17795ff45b94ff772f8e1bec2808f3dbc162c383a0cd0a41"
     private const val _ARCHIVE_FILE_NAME_GEOIP = "geoips.tar.gz"
 
     internal actual val ARCHIVE_SHA256_VALUE_GEOIP: String get() = _ARCHIVE_SHA256_VALUE_GEOIP
     internal actual val ARCHIVE_FILE_NAME_GEOIP: String get() = _ARCHIVE_FILE_NAME_GEOIP
 
-    // TODO: Update with gunzipped archive sha256sum values
-    private const val _ARCHIVE_SHA256_LINUX_X64 = ""
-    private const val _ARCHIVE_SHA256_LINUX_X86 = ""
-    private const val _ARCHIVE_SHA256_MACOS_X64 = ""
-    private const val _ARCHIVE_SHA256_MINGW_X64 = ""
-    private const val _ARCHIVE_SHA256_MINGW_X86 = ""
+    private const val _ARCHIVE_SHA256_VALUE_LINUX_X64 = "787ec03cdb96a328df894c89dddb6de8f616b7d8803b45e293dde0f9fe870bb1"
+    private const val _ARCHIVE_SHA256_VALUE_LINUX_X86 = "7ed8d8887600407a2285619a63a3edd4b9c2136d2c1de0dbc69cd876df348918"
+    private const val _ARCHIVE_SHA256_VALUE_MACOS_X64 = "5b9091852ccf55d5a13e61da74f882b2e4951a5317a0fa9677f2db74b536f24d"
+    private const val _ARCHIVE_SHA256_VALUE_MINGW_X64 = "462e9042a7ad1f9c0f1a7d113b0a8ca3049e54a7df34647a4e96456b1ff77006"
+    private const val _ARCHIVE_SHA256_VALUE_MINGW_X86 = "7ac128a2a92b691bfa924007d220017bb63de5d6ce2dae843351db4613815543"
 
-    internal actual val ARCHIVE_SHA256_LINUX_X64: String get() = _ARCHIVE_SHA256_LINUX_X64
-    internal actual val ARCHIVE_SHA256_LINUX_X86: String get() = _ARCHIVE_SHA256_LINUX_X86
-    internal actual val ARCHIVE_SHA256_MACOS_X64: String get() = _ARCHIVE_SHA256_MACOS_X64
-    internal actual val ARCHIVE_SHA256_MINGW_X64: String get() = _ARCHIVE_SHA256_MINGW_X64
-    internal actual val ARCHIVE_SHA256_MINGW_X86: String get() = _ARCHIVE_SHA256_MINGW_X86
+    internal actual val ARCHIVE_SHA256_VALUE_LINUX_X64: String get() = _ARCHIVE_SHA256_VALUE_LINUX_X64
+    internal actual val ARCHIVE_SHA256_VALUE_LINUX_X86: String get() = _ARCHIVE_SHA256_VALUE_LINUX_X86
+    internal actual val ARCHIVE_SHA256_VALUE_MACOS_X64: String get() = _ARCHIVE_SHA256_VALUE_MACOS_X64
+    internal actual val ARCHIVE_SHA256_VALUE_MINGW_X64: String get() = _ARCHIVE_SHA256_VALUE_MINGW_X64
+    internal actual val ARCHIVE_SHA256_VALUE_MINGW_X86: String get() = _ARCHIVE_SHA256_VALUE_MINGW_X86
 
     private const val _ARCHIVE_FILE_NAME_KMPTOR = "kmptor.tar.gz"
