+++ library/kmp-tor-binary-extract/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/binary/extract/internal/-PlatformConstants.kt
@@ -17,7 +17,7 @@ package io.matthewnelson.kmp.tor.binary.extract.internal
 
 @Suppress("ObjectPropertyName", "SpellCheckingInspection")
 internal actual object PlatformConstants {
-    private const val _ARCHIVE_SHA256_VALUE_GEOIP = "2dce16ea6a5ae3e82c90a1b82bb5ff2eb2a91be98ec2ed539ad722c246752705"
+    private const val _ARCHIVE_SHA256_VALUE_GEOIP = "f3c31f3208a239d030e5367ece4772983c964887f9ba5a34667f3e6e24d94ec2"
     private const val _ARCHIVE_FILE_NAME_GEOIP = "geoips.zip"
 
     @JvmStatic
@@ -25,22 +25,22 @@ internal actual object PlatformConstants {
     @JvmStatic
     internal actual val ARCHIVE_FILE_NAME_GEOIP: String get() = _ARCHIVE_FILE_NAME_GEOIP
 
-    private const val _ARCHIVE_SHA256_LINUX_X64 = "5003d1866d9718d6d2a3c523baf3d91d0a66abc3c75e4ad872ff6d8a879bc6e0"
-    private const val _ARCHIVE_SHA256_LINUX_X86 = "18e1ec895bbb19a59b7ccb73266cafed8d852e19dfc8de0164ccac2642a410fa"
-    private const val _ARCHIVE_SHA256_MACOS_X64 = "23b24fe40c4294b5dfadadca1d13a2950acc08feddbd97bf06cbe4603f64558d"
-    private const val _ARCHIVE_SHA256_MINGW_X64 = "46e195b4e9bceafa5e0ffa4eace400368439b03af6e0718ac3cf265b5b2ac871"
-    private const val _ARCHIVE_SHA256_MINGW_X86 = "dfa4f7c8aac4e50a5646606ac81b3e33b55380b2acb3d9db3dab2530d80249cd"
+    private const val _ARCHIVE_SHA256_VALUE_LINUX_X64 = "7c23de9cdb4647770d45a613f94743fd59fc3eeebbff239ea098fe2a2a2904f6"
+    private const val _ARCHIVE_SHA256_VALUE_LINUX_X86 = "b0af5ee8d488d69badc31c91c88ceb7ec1ca90d77f6e4689e46fafb7885e5ff2"
+    private const val _ARCHIVE_SHA256_VALUE_MACOS_X64 = "63b97fc87f94b56e9f1411ad585fe2a66871b046d5e9d7f3af4672112d851ea2"
+    private const val _ARCHIVE_SHA256_VALUE_MINGW_X64 = "791d22886255261f26a6036a6b011cfcaaeaca9c407f610ac299b389a2049b20"
+    private const val _ARCHIVE_SHA256_VALUE_MINGW_X86 = "6be52087bdedf6f4e0a86fbf23c13a19ee41e6a090e9bc2d097e11a6a41a1367"
 
     @JvmStatic
-    internal actual val ARCHIVE_SHA256_LINUX_X64: String get() = _ARCHIVE_SHA256_LINUX_X64
+    internal actual val ARCHIVE_SHA256_VALUE_LINUX_X64: String get() = _ARCHIVE_SHA256_VALUE_LINUX_X64
     @JvmStatic
-    internal actual val ARCHIVE_SHA256_LINUX_X86: String get() = _ARCHIVE_SHA256_LINUX_X86
+    internal actual val ARCHIVE_SHA256_VALUE_LINUX_X86: String get() = _ARCHIVE_SHA256_VALUE_LINUX_X86
     @JvmStatic
-    internal actual val ARCHIVE_SHA256_MACOS_X64: String get() = _ARCHIVE_SHA256_MACOS_X64
+    internal actual val ARCHIVE_SHA256_VALUE_MACOS_X64: String get() = _ARCHIVE_SHA256_VALUE_MACOS_X64
     @JvmStatic
-    internal actual val ARCHIVE_SHA256_MINGW_X64: String get() = _ARCHIVE_SHA256_MINGW_X64
+    internal actual val ARCHIVE_SHA256_VALUE_MINGW_X64: String get() = _ARCHIVE_SHA256_VALUE_MINGW_X64
     @JvmStatic
-    internal actual val ARCHIVE_SHA256_MINGW_X86: String get() = _ARCHIVE_SHA256_MINGW_X86
+    internal actual val ARCHIVE_SHA256_VALUE_MINGW_X86: String get() = _ARCHIVE_SHA256_VALUE_MINGW_X86
 
     private const val _ARCHIVE_FILE_NAME_KMPTOR = "kmptor.zip"
